requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
Django==5.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
exceptiongroup==1.2.0
fontawesomefree==6.5.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest-cov==4.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240125
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.1.0
